refactor: yup schema
